% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_objects.R
\name{Checksum}
\alias{Checksum}
\title{Checksum Object}
\usage{
Checksum(sha256 = NULL)
}
\arguments{
\item{sha256}{The SHA256 hash of the client state; that is, of the sorted list of all hashes present in the database}
}
\value{
Checksum object
}
\description{
Checksum Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The expected state of a client's local database.
}

