% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_objects.R
\name{ClientInfo}
\alias{ClientInfo}
\title{ClientInfo Object}
\usage{
ClientInfo(clientId = NULL, clientVersion = NULL)
}
\arguments{
\item{clientId}{A client ID that (hopefully) uniquely identifies the client implementation of the Safe Browsing API}

\item{clientVersion}{The version of the client implementation}
}
\value{
ClientInfo object
}
\description{
ClientInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The client metadata associated with Safe Browsing API requests.
}
