% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_objects.R
\name{FetchThreatListUpdatesRequest}
\alias{FetchThreatListUpdatesRequest}
\title{FetchThreatListUpdatesRequest Object}
\usage{
FetchThreatListUpdatesRequest(client = NULL, listUpdateRequests = NULL)
}
\arguments{
\item{client}{The client metadata}

\item{listUpdateRequests}{The requested threat list updates}
}
\value{
FetchThreatListUpdatesRequest object
}
\description{
FetchThreatListUpdatesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes a Safe Browsing API update request. Clients can request updates for multiple lists in a single request. NOTE: Field index 2 is unused. NEXT: 4
}
\seealso{
Other FetchThreatListUpdatesRequest functions: \code{\link{threatListUpdates.fetch}}
}

