% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_objects.R
\name{FetchThreatListUpdatesResponse}
\alias{FetchThreatListUpdatesResponse}
\title{FetchThreatListUpdatesResponse Object}
\usage{
FetchThreatListUpdatesResponse(listUpdateResponses = NULL,
  minimumWaitDuration = NULL)
}
\arguments{
\item{listUpdateResponses}{The list updates requested by the clients}

\item{minimumWaitDuration}{The minimum duration the client must wait before issuing any update request}
}
\value{
FetchThreatListUpdatesResponse object
}
\description{
FetchThreatListUpdatesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
