% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_objects.R
\name{FindFullHashesRequest}
\alias{FindFullHashesRequest}
\title{FindFullHashesRequest Object}
\usage{
FindFullHashesRequest(client = NULL, clientStates = NULL,
  threatInfo = NULL)
}
\arguments{
\item{client}{The client metadata}

\item{clientStates}{The current client states for each of the client's local threat lists}

\item{threatInfo}{The lists and hashes to be checked}
}
\value{
FindFullHashesRequest object
}
\description{
FindFullHashesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to return full hashes matched by the provided hash prefixes.
}
\seealso{
Other FindFullHashesRequest functions: \code{\link{fullHashes.find}}
}
