% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_objects.R
\name{FindFullHashesResponse}
\alias{FindFullHashesResponse}
\title{FindFullHashesResponse Object}
\usage{
FindFullHashesResponse(matches = NULL, minimumWaitDuration = NULL,
  negativeCacheDuration = NULL)
}
\arguments{
\item{matches}{The full hashes that matched the requested prefixes}

\item{minimumWaitDuration}{The minimum duration the client must wait before issuing any find hashes request}

\item{negativeCacheDuration}{For requested entities that did not match the threat list, how long to cache the response}
}
\value{
FindFullHashesResponse object
}
\description{
FindFullHashesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

