% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_objects.R
\name{FindThreatMatchesRequest}
\alias{FindThreatMatchesRequest}
\title{Safe Browsing APIs Objects 
Enables client applications to check web resources (most commonly URLs) against Google-generated lists of unsafe web resources.}
\usage{
FindThreatMatchesRequest(client = NULL, threatInfo = NULL)
}
\arguments{
\item{client}{The client metadata}

\item{threatInfo}{The lists and entries to be checked for matches}
}
\value{
FindThreatMatchesRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:58:29
filename: /Users/mark/dev/R/autoGoogleAPI/googlesafebrowsingv4.auto/R/safebrowsing_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
FindThreatMatchesRequest Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to check entries against lists.
}
\seealso{
Other FindThreatMatchesRequest functions: \code{\link{threatMatches.find}}
}
