% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_objects.R
\name{FindThreatMatchesResponse}
\alias{FindThreatMatchesResponse}
\title{FindThreatMatchesResponse Object}
\usage{
FindThreatMatchesResponse(matches = NULL)
}
\arguments{
\item{matches}{The threat list matches}
}
\value{
FindThreatMatchesResponse object
}
\description{
FindThreatMatchesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

