% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_objects.R
\name{ListThreatListsResponse}
\alias{ListThreatListsResponse}
\title{ListThreatListsResponse Object}
\usage{
ListThreatListsResponse(threatLists = NULL)
}
\arguments{
\item{threatLists}{The lists available for download by the client}
}
\value{
ListThreatListsResponse object
}
\description{
ListThreatListsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

