% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_objects.R
\name{ListUpdateRequest}
\alias{ListUpdateRequest}
\title{ListUpdateRequest Object}
\usage{
ListUpdateRequest(threatType = NULL, platformType = NULL,
  threatEntryType = NULL, state = NULL, constraints = NULL)
}
\arguments{
\item{threatType}{The type of threat posed by entries present in the list}

\item{platformType}{The type of platform at risk by entries present in the list}

\item{threatEntryType}{The types of entries present in the list}

\item{state}{The current state of the client for the requested list (the encrypted client state that was received from the last successful list update)}

\item{constraints}{The constraints associated with this request}
}
\value{
ListUpdateRequest object
}
\description{
ListUpdateRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single list update request.
}
