% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_objects.R
\name{ListUpdateResponse}
\alias{ListUpdateResponse}
\title{ListUpdateResponse Object}
\usage{
ListUpdateResponse(threatType = NULL, threatEntryType = NULL,
  platformType = NULL, responseType = NULL, additions = NULL,
  removals = NULL, newClientState = NULL, checksum = NULL)
}
\arguments{
\item{threatType}{The threat type for which data is returned}

\item{threatEntryType}{The format of the threats}

\item{platformType}{The platform type for which data is returned}

\item{responseType}{The type of response}

\item{additions}{A set of entries to add to a local threat type's list}

\item{removals}{A set of entries to remove from a local threat type's list}

\item{newClientState}{The new client state, in encrypted format}

\item{checksum}{The expected SHA256 hash of the client state; that is, of the sorted list of all hashes present in the database after applying the provided update}
}
\value{
ListUpdateResponse object
}
\description{
ListUpdateResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An update to an individual list.
}
