% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_objects.R
\name{MetadataEntry}
\alias{MetadataEntry}
\title{MetadataEntry Object}
\usage{
MetadataEntry(key = NULL, value = NULL)
}
\arguments{
\item{key}{The metadata entry key}

\item{value}{The metadata entry value}
}
\value{
MetadataEntry object
}
\description{
MetadataEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single metadata entry.
}
