% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_objects.R
\name{RawHashes}
\alias{RawHashes}
\title{RawHashes Object}
\usage{
RawHashes(prefixSize = NULL, rawHashes = NULL)
}
\arguments{
\item{prefixSize}{The number of bytes for each prefix encoded below}

\item{rawHashes}{The hashes, all concatenated into one long string}
}
\value{
RawHashes object
}
\description{
RawHashes Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The uncompressed threat entries in hash format of a particular prefix length. Hashes can be anywhere from 4 to 32 bytes in size. A large majority are 4 bytes, but some hashes are lengthened if they collide with the hash of a popular URL. Used for sending ThreatEntrySet to clients that do not support compression, or when sending non-4-byte hashes to clients that do support compression.
}
