% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_objects.R
\name{RawIndices}
\alias{RawIndices}
\title{RawIndices Object}
\usage{
RawIndices(indices = NULL)
}
\arguments{
\item{indices}{The indices to remove from a lexicographically-sorted local list}
}
\value{
RawIndices object
}
\description{
RawIndices Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A set of raw indices to remove from a local list.
}

