% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_objects.R
\name{RiceDeltaEncoding}
\alias{RiceDeltaEncoding}
\title{RiceDeltaEncoding Object}
\usage{
RiceDeltaEncoding(firstValue = NULL, riceParameter = NULL,
  numEntries = NULL, encodedData = NULL)
}
\arguments{
\item{firstValue}{The offset of the first entry in the encoded data, or, if only a single integer was encoded, that single integer's value}

\item{riceParameter}{The Golomb-Rice parameter, which is a number between 2 and 28}

\item{numEntries}{The number of entries that are delta encoded in the encoded data}

\item{encodedData}{The encoded deltas that are encoded using the Golomb-Rice coder}
}
\value{
RiceDeltaEncoding object
}
\description{
RiceDeltaEncoding Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The Rice-Golomb encoded data. Used for sending compressed 4-byte hashes or compressed removal indices.
}
