% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_functions.R
\docType{package}
\name{safebrowsing_googleAuthR}
\alias{safebrowsing_googleAuthR}
\alias{safebrowsing_googleAuthR-package}
\title{Safe Browsing APIs
Enables client applications to check web resources (most commonly URLs) against Google-generated lists of unsafe web resources.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:22:47
filename: /Users/mark/dev/R/autoGoogleAPI/googlesafebrowsingv4.auto/R/safebrowsing_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item 
}
}

