% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_objects.R
\name{ThreatEntry}
\alias{ThreatEntry}
\title{ThreatEntry Object}
\usage{
ThreatEntry(hash = NULL, url = NULL, digest = NULL)
}
\arguments{
\item{hash}{A hash prefix, consisting of the most significant 4-32 bytes of a SHA256 hash}

\item{url}{A URL}

\item{digest}{The digest of an executable in SHA256 format}
}
\value{
ThreatEntry object
}
\description{
ThreatEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An individual threat; for example, a malicious URL or its hash representation. Only one of these fields should be set.
}

