% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_objects.R
\name{ThreatEntryMetadata}
\alias{ThreatEntryMetadata}
\title{ThreatEntryMetadata Object}
\usage{
ThreatEntryMetadata(entries = NULL)
}
\arguments{
\item{entries}{The metadata entries}
}
\value{
ThreatEntryMetadata object
}
\description{
ThreatEntryMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The metadata associated with a specific threat entry. The client is expected to know the metadata key/value pairs associated with each threat type.
}
