% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_objects.R
\name{ThreatEntrySet}
\alias{ThreatEntrySet}
\title{ThreatEntrySet Object}
\usage{
ThreatEntrySet(compressionType = NULL, rawHashes = NULL,
  rawIndices = NULL, riceHashes = NULL, riceIndices = NULL)
}
\arguments{
\item{compressionType}{The compression type for the entries in this set}

\item{rawHashes}{The raw SHA256-formatted entries}

\item{rawIndices}{The raw removal indices for a local list}

\item{riceHashes}{The encoded 4-byte prefixes of SHA256-formatted entries, using a Golomb-Rice encoding}

\item{riceIndices}{The encoded local, lexicographically-sorted list indices, using a Golomb-Rice encoding}
}
\value{
ThreatEntrySet object
}
\description{
ThreatEntrySet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A set of threats that should be added or removed from a client's local database.
}
