% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_objects.R
\name{ThreatInfo}
\alias{ThreatInfo}
\title{ThreatInfo Object}
\usage{
ThreatInfo(threatTypes = NULL, platformTypes = NULL,
  threatEntryTypes = NULL, threatEntries = NULL)
}
\arguments{
\item{threatTypes}{The threat types to be checked}

\item{platformTypes}{The platform types to be checked}

\item{threatEntryTypes}{The entry types to be checked}

\item{threatEntries}{The threat entries to be checked}
}
\value{
ThreatInfo object
}
\description{
ThreatInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The information regarding one or more threats that a client submits when checking for matches in threat lists.
}

