% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_objects.R
\name{ThreatListDescriptor}
\alias{ThreatListDescriptor}
\title{ThreatListDescriptor Object}
\usage{
ThreatListDescriptor(threatType = NULL, platformType = NULL,
  threatEntryType = NULL)
}
\arguments{
\item{threatType}{The threat type posed by the list's entries}

\item{platformType}{The platform type targeted by the list's entries}

\item{threatEntryType}{The entry types contained in the list}
}
\value{
ThreatListDescriptor object
}
\description{
ThreatListDescriptor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes an individual threat list. A list is defined by three parameters: the type of threat posed, the type of platform targeted by the threat, and the type of entries in the list.
}
