% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_functions.R
\name{threatListUpdates.fetch}
\alias{threatListUpdates.fetch}
\title{Fetches the most recent threat list updates. A client can request updates for multiple lists at once.}
\usage{
threatListUpdates.fetch(FetchThreatListUpdatesRequest)
}
\arguments{
\item{FetchThreatListUpdatesRequest}{The \link{FetchThreatListUpdatesRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/safe-browsing/}{Google Documentation}

Other FetchThreatListUpdatesRequest functions: \code{\link{FetchThreatListUpdatesRequest}}
}

