% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_objects.R
\name{ThreatMatch}
\alias{ThreatMatch}
\title{ThreatMatch Object}
\usage{
ThreatMatch(threatType = NULL, platformType = NULL,
  threatEntryType = NULL, threat = NULL, threatEntryMetadata = NULL,
  cacheDuration = NULL)
}
\arguments{
\item{threatType}{The threat type matching this threat}

\item{platformType}{The platform type matching this threat}

\item{threatEntryType}{The threat entry type matching this threat}

\item{threat}{The threat matching this threat}

\item{threatEntryMetadata}{Optional metadata associated with this threat}

\item{cacheDuration}{The cache lifetime for the returned match}
}
\value{
ThreatMatch object
}
\description{
ThreatMatch Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A match when checking a threat entry in the Safe Browsing threat lists.
}
