% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safebrowsing_functions.R
\name{threatMatches.find}
\alias{threatMatches.find}
\title{Finds the threat entries that match the Safe Browsing lists.}
\usage{
threatMatches.find(FindThreatMatchesRequest)
}
\arguments{
\item{FindThreatMatchesRequest}{The \link{FindThreatMatchesRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/safe-browsing/}{Google Documentation}

Other FindThreatMatchesRequest functions: \code{\link{FindThreatMatchesRequest}}
}

